ProxyTestBase class
